Fixing hoisting issue.
